+++ src/Tea-Shop.Application/Users/Commands/DeleteUserCommand/DeleteUserHandler.cs
@@ -1,6 +1,8 @@
-﻿using CSharpFunctionalExtensions;
+﻿using System.Data;
+using CSharpFunctionalExtensions;
 using Microsoft.Extensions.Logging;
 using Tea_Shop.Application.Abstractions;
+using Tea_Shop.Application.Database;
 using Tea_Shop.Contract.Users;
 using Tea_Shop.Domain.Users;
 using Tea_Shop.Shared;
@@ -12,13 +14,16 @@ public class DeleteUserHandler:
 {
     private readonly IUsersRepository _usersRepository;
     private readonly ILogger<DeleteUserHandler> _logger;
+    private readonly ITransactionManager _transactionManager;
 
     public DeleteUserHandler(
         IUsersRepository usersRepository,
-        ILogger<DeleteUserHandler> logger)
+        ILogger<DeleteUserHandler> logger,
+        ITransactionManager transactionManager)
     {
         _usersRepository = usersRepository;
         _logger = logger;
+        _transactionManager =  transactionManager;
     }
 
     public async Task<Result<UserWithOnlyIdDto, Error>> Handle(
@@ -26,6 +31,20 @@ public class DeleteUserHandler:
         CancellationToken cancellationToken)
     {
         _logger.LogDebug("Handling {handler}", nameof(DeleteUserHandler));
+
+
+        var transactionScopeResult = await _transactionManager.BeginTransactionAsync(
+            IsolationLevel.RepeatableRead,
+            cancellationToken);
+
+        if (transactionScopeResult.IsFailure)
+        {
+            _logger.LogError("Failed to begin transaction while deleting user");
+            return transactionScopeResult.Error;
+        }
+
+        using var transactionScope = transactionScopeResult.Value;
+
 
         var deleteResult = await _usersRepository.DeleteUser(
             new UserId(command.Request.UserId),
@@ -37,10 +56,25 @@ public class DeleteUserHandler:
                 "Failed to delete user {productId}",
                 command.Request.UserId);
 
+            transactionScope.Rollback();
+
             return deleteResult.Error;
         }
 
-        _logger.LogInformation(
+
+        await _transactionManager.SaveChangesAsync(cancellationToken);
+
+        var commitedResult = transactionScope.Commit();
+
+        if (commitedResult.IsFailure)
+        {
+            _logger.LogError("Failed to commit result while deleting user");
+            transactionScope.Rollback();
+            return commitedResult.Error;
+        }
+
+
+        _logger.LogDebug(
             "User with id {UserId} deleted.",
             command.Request.UserId);
